requirements.txt
streamlit==1.50.0
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
Pillow
numpy
scikit-image
lpips
pytorch_fid
pytorch-fid
